LSモジュールの削除
