+++ 08.ls_object/test/format_test.rb
@@ -1,13 +1,18 @@
 # frozen_string_literal: true
 
 require 'minitest/autorun'
-require_relative '../lib/ls'
-require_relative '../lib/ls/format'
+require_relative '../lib/file_stat'
+require_relative '../lib/format'
 
 class FormatTest < Minitest::Test
+  def generate_file_stats_classes(path)
+    paths = Dir.foreach(path).map { |entry| File.join(path, entry) }
+    paths.map { |p| FileStat.new(p) }
+  end
+
   def test_default_format
-    file_stats = LS.generate_file_stats_class('./08.ls_object/testdata')
-    format = LS::Format.new(file_stats)
+    file_stats = generate_file_stats_classes('./08.ls_object/testdata')
+    format = Format.new(file_stats)
     actual = format.render
 
     expected = <<~TEXT.chomp
@@ -17,8 +22,8 @@ class FormatTest < Minitest::Test
   end
 
   def test_all_file_format
-    file_stats = LS.generate_file_stats_class('./08.ls_object/testdata')
-    format = LS::Format.new(file_stats, all: true)
+    file_stats = generate_file_stats_classes('./08.ls_object/testdata')
+    format = Format.new(file_stats, all: true)
     actual = format.render
 
     expected = <<~TEXT.chomp
@@ -29,8 +34,8 @@ class FormatTest < Minitest::Test
   end
 
   def test_reverse_format
-    file_stats = LS.generate_file_stats_class('./08.ls_object/testdata/testdir')
-    format = LS::Format.new(file_stats, reverse: true)
+    file_stats = generate_file_stats_classes('./08.ls_object/testdata/testdir')
+    format = Format.new(file_stats, reverse: true)
     actual = format.render
 
     expected = <<~TEXT.chomp
@@ -40,33 +45,33 @@ class FormatTest < Minitest::Test
   end
 
   def test_all_long_listing_format
-    file_stats = LS.generate_file_stats_class('./08.ls_object/testdata/testdir')
-    format = LS::Format.new(file_stats, all: true, long: true)
+    file_stats = generate_file_stats_classes('./08.ls_object/testdata/testdir')
+    format = Format.new(file_stats, all: true, long: true)
     actual = format.render
 
     expected = <<~TEXT.chomp
       total 8
-      drwxr-xr-x 2 misono misono 4096 Sep 19 09:09 .
-      drwxr-xr-x 3 misono misono 4096 Sep 19 09:00 ..
-      -rw-r--r-- 1 misono misono    0 Sep 19 08:59 bar.txt
-      -rw-r--r-- 1 misono misono    0 Sep 19 08:59 foo.txt
+      drwxr-xr-x 2 misono misono 4096 Sep 25 12:02 .
+      drwxr-xr-x 3 misono misono 4096 Sep 25 12:02 ..
+      -rw-r--r-- 1 misono misono    0 Sep 25 12:02 bar.txt
+      -rw-r--r-- 1 misono misono    0 Sep 25 12:02 foo.txt
     TEXT
     assert_equal expected, actual
   end
 
   def test_all_long_reverse_format
-    file_stats = LS.generate_file_stats_class('./08.ls_object/testdata')
-    format = LS::Format.new(file_stats, all: true, long: true, reverse: true)
+    file_stats = generate_file_stats_classes('./08.ls_object/testdata')
+    format = Format.new(file_stats, all: true, long: true, reverse: true)
     actual = format.render
 
     expected = <<~TEXT.chomp
       total 16
-      drwxr-xr-x 2 misono misono 4096 Sep 19 09:09 testdir
-      -rw-r--r-- 1 misono misono    0 Sep 19 08:59 .hoge
-      -rw-r--r-- 1 misono misono  446 Sep 19 09:00 foo.md
-      -rw-r--r-- 1 misono misono    0 Sep 19 08:59 bar.txt
-      drwxr-xr-x 5 misono misono 4096 Sep 19 09:11 ..
-      drwxr-xr-x 3 misono misono 4096 Sep 19 09:00 .
+      drwxr-xr-x 2 misono misono 4096 Sep 25 12:02 testdir
+      -rw-r--r-- 1 misono misono    0 Sep 25 12:02 .hoge
+      -rw-r--r-- 1 misono misono  446 Sep 25 12:02 foo.md
+      -rw-r--r-- 1 misono misono    0 Sep 25 12:02 bar.txt
+      drwxr-xr-x 5 misono misono 4096 Sep 25 12:59 ..
+      drwxr-xr-x 3 misono misono 4096 Sep 25 12:02 .
     TEXT
     assert_equal expected, actual
   end
